+++ packages/hono-takibi/src/generators/hono/openapi/route/request/body/generate-request-body.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { generateRequestBody } from './generate-request-body'
-import type { Content } from '../../../types'
+import type { Content } from '../../../../../../types'
 
 const generateRequestBodyTestCases: {
   required: boolean
