+++ packages/hono-takibi/src/generators/hono/openapi/route/request/object/generate-format-request-object.test.ts
